requirements.txt
flask~=2.1.1
ariadne~=0.15.0
SQLAlchemy~=1.4.36
flask-sqlalchemy
flask-cors
face_recognition
requests~=2.26.0
gql~=3.2.0
sqlalchemy-migrate
Pillow~=9.1.0